Add Items#Chart; Fix Item#Amount bug
Added the Items#Chart page to show expenditures by tags over time. Fixed
a bug in the Item#Amount field that caused the database to store a
negative amount (i.e., a credit) as positive (i.e., a debit) because the
field's behaviour was modified in the application-level, but the meaning
was not modified in the database. This bug was fixed by restoring
the original functionality in the application-level, but then adding new functionality to a temporary field called `TheRealAmount`.