Use decimal instead of Money
Use the `decimal` type instead of `Money` because the quantity of money
received in year X does not change over time. We can still use the
`Money` type for comparing the purchasing power of money at two
different dates. As in, what is the equivalent of 20 dollars in 1965
dollars in 1995 dollars? So for now we will keep the `Money` type
alive.